requirements.txt
numpy==1.18.3
opencv-python==4.2.0.34
pkg-resources==0.0.0
pyfakewebcam==0.1.0
Pillow==7.1.1
